pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-textanalytics>=5.3.0",
    "fastapi>=0.129.0",
    "nltk>=3.9.2",
    "scikit-learn==1.7.2",
    "joblib==1.5.2",
    "numpy>=2.4.2",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "pandas>=3.0.1",
    "pyodbc>=5.3.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "ruff>=0.15.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
